Add in CCAL library for the PrimEx CompCal ("Compton Calorimeter").
The geometry and mods to HDGeant were done by Pawel A. last summer. These changes bring the values into DANA in a way that is consistent with how the FCAL is done. This makes the DCCALHit and DCCALHit:TRUTH objects available (the former only after running mcsmear). Also, the DCCALTruthShower objects are now available to DANA programs.

No reconstruction of CCAL data is included yet.